set game data id to BGG id instead of search id
When performing a Thing search using id parameter, the id doesn't match the one that come from the BGG's XMLAPI2.

Way to reproduce: perform a game_list search using id = 1 as a parameter, the response item has an id of 2.

Found the id mismatch while coding a program to get the game expansions from a game after having got the whole data. The expansion ids didn't match the correct items.